minor bug

+++ RICS_Qt/mainwindow.cpp
@@ -164,10 +164,10 @@ void MainWindow::read_settings(){
 
     arm_movement_degrees = settings.value("arm_movement_degrees", 5).toInt();
     QHoverSensitiveButton::hoverTime = settings.value("hoverTime", 3000).toInt();
-    fetch_x = settings.value("fetch_x", 10);
-    fetch_y = settings.value("fetch_y", 0);
-    fetch_z = settings.value("fetch_z", 0);
-    fetch_claw = settings.value("fetch_claw", 0);
+    fetch_x = settings.value("fetch_x", 10).toInt();
+    fetch_y = settings.value("fetch_y", 0).toInt();
+    fetch_z = settings.value("fetch_z", 0).toInt();
+    fetch_claw = settings.value("fetch_claw", 0).toInt();
 
 }
